+++ plugins/modules/na_ontap_debug.py
@@ -3,7 +3,7 @@
 create Debug module to diagnose netapp-lib import and connection
 """
 
-# (c) 2020-2025, NetApp, Inc
+# (c) 2020-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
@@ -29,6 +29,31 @@ options:
     - The vserver name to test for ZAPI tunneling.
     required: false
     type: str
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See README for example usage.
 '''
 EXAMPLES = """
 - name: Check import netapp-lib
@@ -36,6 +61,13 @@ EXAMPLES = """
     hostname: "{{ netapp_hostname }}"
     username: "{{ netapp_username }}"
     password: "{{ netapp_password }}"
+
+- name: Debug vserver
+  netapp.ontap.na_ontap_debug:
+    hostname: "{{ netapp_hostname }}"
+    username: "{{ netapp_username }}"
+    password: "{{ netapp_password }}"
+    vserver: ansibleSVM
 """
 
 RETURN = """
@@ -57,12 +89,21 @@ class NetAppONTAPDebug(object):
         self.argument_spec.update(dict(
             vserver=dict(required=False, type="str"),
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
-            argument_spec=self.argument_spec
+            argument_spec=self.argument_spec,
+            required_if=[
+                ('use_lambda', True, ['lambda_config']),
+            ],
         )
         self.na_helper = NetAppModule()
         self.parameters = self.na_helper.set_parameters(self.module.params)
         self.rest_api = netapp_utils.OntapRestAPI(self.module)
+        self.use_rest = self.rest_api.is_rest()
+        if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
         self.log_list = []
         self.error_list = []
         self.note_list = []
@@ -222,8 +263,9 @@ class NetAppONTAPDebug(object):
         # report Ansible and our collection versions
         self.list_versions()
 
-        # check import netapp-lib
-        self.import_lib()
+        # check import netapp-lib when not using REST
+        if not self.use_rest:
+            self.import_lib()
 
         # check zapi connection errors only if import successful
         if netapp_utils.has_netapp_lib():