+++ plugins/modules/na_ontap_rest_info.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2020-2025, NetApp, Inc
+# (c) 2020-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 """ NetApp ONTAP Info using REST APIs """
@@ -35,6 +35,7 @@ notes:
   - I(role_info) there is not REST equivalent.
   - I(subsys_health_info) there is not REST equivalent.
   - I(volume_move_target_aggr_info) there is not REST equivalent.
+  - Supports AWS Lambda proxy functionality. See README for example usage.
 
 options:
   state:
@@ -319,6 +320,27 @@ options:
       - if false, HAL-encoded links are disabled in the REST calls.
     default: true
     type: bool
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
 '''
 
 EXAMPLES = '''
@@ -474,9 +496,13 @@ class NetAppONTAPGatherInfo(object):
             ignore_api_errors=dict(type='list', elements='str', required=False),
             hal_linking=dict(required=False, type='bool', default=True),
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
 
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
+            required_if=[
+                ('use_lambda', True, ['lambda_config']),
+            ],
             supports_check_mode=True
         )